Fixed tactile functions

+++ ur_robot_driver/include/inspire_hand/hand_control_lib_v2.h
@@ -74,9 +74,9 @@ public:
     int getRegisterValue(int reg_addr) {
         return readRegister(reg_addr); // 通过公有方法调用私有方法
     }
-    cv::Mat convert_tactile_data_to_image(const std::vector<std::vector<std::vector<int>>>& multi_tactile_data, int rows=256, int cols=256);
-    std::vector<std::vector<int>> resize_tactile_data(const std::vector<int>& v, int rows, int cols);
-    bool read_tactile(int start_addr, std::vector<int>& tactile_data, int num_values);
+    cv::Mat convert_tactile_data_to_image(const std::vector<std::vector<std::vector<uint16_t>>>& multi_tactile_data, int rows=256, int cols=256);
+    std::vector<std::vector<uint16_t>> resize_tactile_data(const std::vector<uint16_t>& v, int rows, int cols);
+    bool read_tactile(int start_addr, std::vector<uint16_t>& tactile_data, int num_values);
     bool get_tactile_data();
 
     //hand state variables              
@@ -93,7 +93,7 @@ public:
     double setangle_[6];
     double setangle_cmd_[6];
     double setforce_[6];
-    std::vector<std::vector<std::vector<int>>> multi_tactile_data_;
+    std::vector<std::vector<std::vector<uint16_t>>> multi_tactile_data_;
     cv::Mat multi_tactile_image_;
 
 private:
